refactor(mobile): v1.35.0 PR-E — split mobile.js + help-doctor dir-module coverage
Split lib/commands/mobile.js (1220 lines) into a mobile/ directory-module
mirroring the existing icerik/ pattern, and teach help-doctor to audit
directory-modules so split commands stay covered.

Mobile split (verbatim move — CLI grammar + all output strings unchanged):
- mobile/{init,version,build,release,assets,crash,deeplink,ota}.js (one per
  subcommand) + index.js (runMobile dispatch) + help.js (mobileHelp).
- bin/badi.js imports ../lib/commands/mobile/index.js.

help-doctor directory-module coverage (re-scoped from PR-A):
- runHelpDoctor + auditFiles now discover dirs containing index.js and audit
  the index.js dispatch combined with help.js help text. New exported
  detectDriftSource(source, opts). This RESTORES coverage silently dropped when
  icerik/ and plugin/ were split, and covers the new mobile/: 35 -> 38 audited
  commands, all drift-free (no allowlist entries needed).

Tests 1315 -> 1317 (detectDriftSource + dir-module auditFiles; the 33 mobile
e2e tests pass unchanged). biome clean. release check green. Freeze-compatible.
