+++ lib/commands/doctor.js
@@ -1,4 +1,4 @@
-import { readdirSync } from "node:fs";
+import { existsSync, readdirSync } from "node:fs";
 import { join, resolve } from "node:path";
 import { chalk, showBanner } from "../cli.js";
 import { HARNESSES, resolveHarnesses } from "../harnesses/index.js";
@@ -43,9 +43,16 @@ function runHelpDoctor(args) {
 
 	let files;
 	try {
-		files = readdirSync(commandsDir)
-			.filter((f) => f.endsWith(".js"))
-			.map((f) => join(commandsDir, f));
+		// Include both top-level *.js command files AND directory-modules
+		// (a dir containing index.js, e.g. mobile/, icerik/) so split commands
+		// stay under help-doctor coverage.
+		files = readdirSync(commandsDir, { withFileTypes: true })
+			.filter((d) =>
+				d.isDirectory()
+					? existsSync(join(commandsDir, d.name, "index.js"))
+					: d.name.endsWith(".js"),
+			)
+			.map((d) => join(commandsDir, d.name));
 	} catch {
 		console.error(chalk.red(`lib/commands/ not found: ${commandsDir}`));
 		console.error(
