+++ lib/help-doctor.js
@@ -17,7 +17,8 @@
 //   - When the help body is split into an external function (showHelp), the whole file is the body
 // Allowlist mechanism: per-file exceptions via .claude/help-doctor.allow.json.
 
-import { existsSync, readFileSync } from "node:fs";
+import { existsSync, readFileSync, statSync } from "node:fs";
+import { join } from "node:path";
 
 // Capture strings that appear as subcommands on the parser side.
 const SUBCOMMAND_PATTERNS = [
@@ -108,7 +109,19 @@ function extractHelpBody(source) {
  * @returns {{ missingSubs: string[], missingFlags: string[] }}
  */
 export function detectDrift(filePath, opts = {}) {
-	const source = readFileSync(filePath, "utf-8");
+	return detectDriftSource(readFileSync(filePath, "utf-8"), opts);
+}
+
+/**
+ * Drift detection over a raw source string (file content already read). Lets
+ * directory-modules be audited by combining their index.js dispatch with their
+ * help.js help text — see auditFiles.
+ *
+ * @param {string} source
+ * @param {object} [opts]
+ * @returns {{ missingSubs: string[], missingFlags: string[] }}
+ */
+export function detectDriftSource(source, opts = {}) {
 	const parserSubs = extractParserSubcommands(source);
 	const parserFlags = extractParserFlags(source);
 	const helpBody = extractHelpBody(source);
@@ -163,7 +176,19 @@ export function auditFiles(filePaths, opts = {}) {
 		const entry = allowlist[fileName] || {};
 		const allowSubs = new Set(entry.subs || []);
 		const allowFlags = new Set(entry.flags || []);
-		const drift = detectDrift(filePath, { allowSubs, allowFlags });
+		// Directory-module (e.g. mobile/, icerik/): audit the index.js dispatch
+		// against the help.js help text combined, so split commands stay covered
+		// (a plain .js filter would silently drop them from help-doctor).
+		let drift;
+		if (existsSync(filePath) && statSync(filePath).isDirectory()) {
+			const idx = join(filePath, "index.js");
+			const help = join(filePath, "help.js");
+			let source = existsSync(idx) ? readFileSync(idx, "utf-8") : "";
+			if (existsSync(help)) source += `\n${readFileSync(help, "utf-8")}`;
+			drift = detectDriftSource(source, { allowSubs, allowFlags });
+		} else {
+			drift = detectDrift(filePath, { allowSubs, allowFlags });
+		}
 		if (drift.missingSubs.length > 0 || drift.missingFlags.length > 0) {
 			results.push({ file: fileName, ...drift });
 		}
